chg: dev: Avoid error if key in table of face indices for sub meshes already exists.
When we ran test exports the package was throwing an error here and messed up the exported result. We still need to do more research on why this is happening. Maybe we need to update the existing entry or even have to flush the entire dictionary at some point. This definitely works as a quick fix.